lambda/requirements.txt
python-dotenv
langchain 
langgraph
streamlit
langchain-community
langchain-aws
boto3
python-multipart
jinja2
graphviz
mistralai

